Switch index format from packed u64 to triples
Was: [nblocks] storing u64( (ni<<32)|(hi<<16) | wi)]
Now: [nblocks, 3] storing (ni,hi,wi)

+++ sbnet_tensorflow/sbnet_ops/reduce_mask.cu
@@ -48,13 +48,15 @@ template <typename T> struct ReduceMaskFunctor<GPUDevice, T> {
         int bStrW,                      // Block stride, width.
         int bCntH,                      // Number of blocks, height.
         int bCntW,                      // Number of blocks, width.
-        unsigned int numBins,
-        unsigned int binSize,
-        long long* activeBlockIndices,  // Indices of active blocks.
+        unsigned int numBins,           // number of bins in binCounts
+        unsigned int binSize,           // maximum size of each counter bin
+        int* activeBlockIndices,        // triples of [n, ih, iw] indices for active blocks.
         int* binCounts,                 // Number of indices of active blocks.
-        bool avgPool
+        bool avgPool                    // true for avg pooling, false for max pooling
         )
     {
+        // TODO
+        // We can do better here in terms of grid/block partitioning but this is not currently a perf bottleneck
         zeroBlockCounters<<<24, 1024, 0, d.stream()>>>(numBins, (unsigned int*) binCounts);
         dim3 block(std::min(DIVUP(bSzH*bSzW, 32)*32, 1024), 1, 1);
         dim3 grid(bCntW, bCntH, N);
@@ -63,14 +65,13 @@ template <typename T> struct ReduceMaskFunctor<GPUDevice, T> {
             numBins,   // number of bins to partition activeBlockIndices to reduce atomics pressure
             binSize,
             (unsigned int*) binCounts, // counts for sub-blocks, initialized to 0
-            (unsigned long long*) activeBlockIndices,
+            (int*) activeBlockIndices,
             bOffsH0,
             bOffsW0,      // generally negative - first block element offset for correct padding
             bSzH, bSzW,   // block sizes
             bStrH, bStrW, // block strides
             bCntH, bCntW, // block counts
-            avgPool
-            );
+            avgPool);
     }
 };
